hyakvnc.py: Use srun instead of ssh to launch process on specific job allocation
This commit handles the dilemma where if there are two jobs allocating
the same node, it was impossible to determine if ssh'ing into the node
is launching on one or the other job. This becomes an issue when killing
a specific vnc job does not kill the vnc session because there was no
association between the two. With `srun --jobid=##### <command> <args>`,
we can clearly make the association between a task and a job. So when we
kill a job, it will kill its associated vnc session.

Example:
```
% ./hyakvnc.py -f
...
=====================
Run the following in a new terminal window:
        ssh -N -f -L 5900:127.0.0.1:5900 hansem7@klone.hyak.uw.edu
then connect to VNC session at localhost:5900
=====================
% ./hyakvnc.py -f
...
=====================
Run the following in a new terminal window:
        ssh -N -f -L 5901:127.0.0.1:5901 hansem7@klone.hyak.uw.edu
then connect to VNC session at localhost:5901
=====================
% ./hyakvnc.py --status
Active ece_vnc jobs:
        Job ID: 1511263
                SubNode: n3030 <---- same SubNode on both jobs
                VNC active: True
                VNC display number: 1
                VNC port: 5901
                Mapped LoginNode port: 5900
                Time left: 3:58:56
                Run command: ssh -N -f -L 5900:127.0.0.1:5900 hansem7@klone.hyak.uw.edu
        Job ID: 1511264
                SubNode: n3030
                VNC active: True
                VNC display number: 2
                VNC port: 5902
                Mapped LoginNode port: 5901
                Time left: 3:59:33
                Run command: ssh -N -f -L 5901:127.0.0.1:5901 hansem7@klone.hyak.uw.edu
% ./hyakvnc.py --kill 1511263
Attempting to kill 1511263
        Canceling job ID 1511263
% ./hyakvnc.py --status
Active ece_vnc jobs:
        Job ID: 1511264
                SubNode: n3030
                VNC active: True <------ VNC session is still alive
                VNC display number: 2
                VNC port: 5902
                Mapped LoginNode port: 5901
                Time left: 3:59:33
                Run command: ssh -N -f -L 5901:127.0.0.1:5901 hansem7@klone.hyak.uw.edu
```