hyakvnc.py: check_auth_keys: return True if contents of pubkey matches
This improves check robustness by checking the existence of public key
contents in ~/.ssh/authorized_keys.

If the word 'klone' exists in authorized_keys, but the public key does
not exist, then login node cannot execute commands on subnode via SSH.

If a key is not added to authorized_keys, hyakvnc.py will ask the user
to add it to continue.